feat: 2초 후 나가기
